% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCDClinks.R
\name{buildCDClinks}
\alias{buildCDClinks}
\title{Build CDC link}
\usage{
buildCDClinks(
  and_terms = "",
  not_terms = "",
  exact_phrase = "",
  or_terms = "",
  date_from = "",
  date_to = "",
  pages = 1,
  start_page = 1,
  language = ""
)
}
\arguments{
\item{and_terms}{Vector of alphanumeric terms searched using the AND Boolean operator,
specified by the CDC as 'with all of the words'.}

\item{not_terms}{Vector of alphanumeric terms searched using the NOT Boolean operator,
specified by the CDC as 'without the words'.}

\item{exact_phrase}{Vector of alphanumeric terms enclosed in inverted commas and searched
as phrases (e.g. "large cat"), specified by the CDC as 'with the exact phrase'.}

\item{or_terms}{Vector of alphanumeric terms searched using the OR Boolean operator,
specified by the CDC as 'with at least one of the words'.}

\item{date_from}{The date (e.g. 01/01/2000) from which searching will be performed
(inclusive). If no value provided, all dates are searched.}

\item{date_to}{The date (e.g. 01/10/2020) to which searching will be performed
(inclusive). If no value provided, all dates are searched.}

\item{pages}{Integer for the number of pages of search results to be returned (one link per
page). A maximum of 100 pages can be displayed in the CDC. The default value is 1.}

\item{start_page}{Integer specifying which page(s) of search results should be displayed.
If multiple pages are selected, multiple URLs are returned, one for each page of ten
search results.}

\item{language}{Two-letter language code for search language. The default is 'en' (English).}
}
\value{
A link containing the specified search results.
}
\description{
Constructs series of CDC website search page URLs
}
\details{
Function takes as input a partial Boolean search string and produces a functioning (set of)
URLs; one for each page og search results on the CDC website.
}
\examples{
and_terms <- c('disease', 'spread')
not_terms <- c('animal', 'fish')
exact_phrase <- c('corona virus')
or_terms <- c('pandemic', 'global')
date_from <- '01/01/1980'
date_to <- '10/11/2020'
link <- buildCDClinks(and_terms = and_terms,
    not_terms = not_terms,
    exact_phrase = exact_phrase,
    or_terms = or_terms,
    date_from = date_from,
    date_to = date_to,
    pages = 3)
link;
}
