% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_htmls.R
\name{save_code}
\alias{save_code}
\title{Download one results page as html file}
\usage{
save_code(url, browser = "firefox")
}
\arguments{
\item{url}{One URLs corresponding to a page of CDC search results.}

\item{browser}{The name of the browser used for scraping. Default is 'firefox'.}
}
\value{
An object containing the HTML code. An html file is also saved to the working directory.
}
\description{
Downloads one page of CDC results from a URLs as html files with
a specific wait-time to avoid IP address blocking.
}
\examples{
url <- 'https://search.cdc.gov/search/index.html?all=disease\%20spread&none=animal\%20fish&exact=corona\%20virus&any=pandemic\%20global&date1=01\%2F01\%2F1980&date2=10\%2F11\%2F2020&dpage=1'
save_html(url, pause = 5, backoff = FALSE);
}
