% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_htmls.R
\name{save_codes}
\alias{save_codes}
\title{Download multiple results pages as html files}
\usage{
save_codes(urls, browser = "firefox")
}
\arguments{
\item{urls}{One or more URLs corresponding to pages of CDC search results.}

\item{browser}{The name of the browser used for scraping. Default is 'firefox'.}
}
\value{
A list of objects containing the HTML codes for each file. html files are also saved to the working directory.
}
\description{
Downloads one or more CDC URLs as html files with a specific wait-time
to avoid IP address blocking.
}
\examples{
urls <- c("https://search.cdc.gov/search/index.html?all=disease\%20spread&none=animal\%20fish&exact=corona\%20virus&any=pandemic\%20global&date1=01\%2F01\%2F1980&date2=10\%2F11\%2F2020&dpage=1#results",
"https://search.cdc.gov/search/index.html?all=disease\%20spread&none=animal\%20fish&exact=corona\%20virus&any=pandemic\%20global&date1=01\%2F01\%2F1980&date2=10\%2F11\%2F2020&dpage=2#results",
"https://search.cdc.gov/search/index.html?all=disease\%20spread&none=animal\%20fish&exact=corona\%20virus&any=pandemic\%20global&date1=01\%2F01\%2F1980&date2=10\%2F11\%2F2020&dpage=3#results")
codes <- save_codes(urls);
}
